backend/requirements/local.txt
-r base.txt

Werkzeug[watchdog]==2.2.0
ipdb==0.13.9
psycopg2==2.9.3
watchfiles==0.16.1

# Testing
# ------------------------------------------------------------------------------
mypy==0.971
django-stubs==1.12.0  
pytest==7.1.2  
pytest-sugar==0.9.5  
djangorestframework-stubs==1.7.0  

# Code quality
# ------------------------------------------------------------------------------
flake8==4.0.1
flake8-isort==4.1.2.post0  
coverage==6.4.2
black==22.6.0
pylint-django==2.5.3  
pylint-celery==0.3  
pre-commit==2.20.0   

# Django
# ------------------------------------------------------------------------------
factory-boy==3.2.1
django-debug-toolbar==3.5.0
django-extensions==3.2.0 
django-coverage-plugin==2.0.3
pytest-django==4.5.2
